feat: move displayFlashMessages function into module
